requirements.txt
python-telegram-bot[job-scheduler]==21.0
google-cloud-bigquery==3.25.0
playwright==1.44.0
